Expansion error when number in pos 2 Fixes #112
